Eslint refactor
Переписал работу вкладок в MoviePage чтобы избавиться от всех предупреждений от Eslint
